requirements_temp.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
SQLAlchemy==2.0.40
cryptography==36.0.2
python-dotenv==1.0.1
qrcode==7.4.2
requests==2.31.0
google-generativeai==0.3.1
pillow==10.2.0
python-escpos==3.0a8
